:hammer: :bug: Fix Texttest runner
We're going to refactor the Java version of this kata. We removed the
references to other languages and enabled java in `config.gr`.
However, I couldn't get the classpath to work when using
`environment.gr`, so I renamed the file to `interpreter_options.gr` as
according to the Texttest documentation.
We explicitely add the `-cp` flag to define our classpath. We can still
use the `${TEXTTEST_HOME}` variable.

We now get a pass on the test.