:hammer: Extract update behavior to dedicated classes
We're going to use polymorphism together with a factory method. We could
have extended `Item` instead, but then the goblin will kill us.

We're not going to extract each and every step into separate methods
which then can be overridden with fancy constructions, which will make
the code unreadable again.